#17 added LoginPanel.kt
added loginPanel, set it as first panel and moved some code to corresponding folder
