+++ composeApp/src/commonMain/kotlin/com/braiso_22/terracambio/chat/presentation/components/imagepreview/ImagePreview.kt
@@ -1,0 +1,5 @@
+package com.braiso_22.terracambio.chat.presentation.components.imagepreview
+
+data class ImagePreview(
+    val text: String,
+)
