--- composeApp/src/commonMain/kotlin/com/braiso_22/terracambio/listing/presentation/chat/components/imagepreview/ImagePreview.kt
@@ -1,5 +1,0 @@
-package com.braiso_22.terracambio.listing.presentation.chat.components.imagepreview
-
-data class ImagePreview(
-    val text: String,
-)
